% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_z.R
\name{categorical_z}
\alias{categorical_z}
\title{Calculate a z-score for categorical data}
\usage{
categorical_z(p_hat, p, n)
}
\arguments{
\item{p_hat}{A sample proportion}

\item{p}{A population proportion}

\item{n}{Sample size}
}
\description{
Calculate a z-score for categorical data
}
