% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_prediction.R
\name{linear_model_prediction}
\alias{linear_model_prediction}
\title{Evaluate a linear model at a given point}
\usage{
linear_model_prediction(model, value)
}
\arguments{
\item{model}{A linear model}

\item{value}{A value to make a prediction with}
}
\description{
Evaluate a linear model at a given point
}
