% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_z.R
\name{numeric_z}
\alias{numeric_z}
\title{Calculate a z-score for numeric data}
\usage{
numeric_z(x, mu, s)
}
\arguments{
\item{x}{A data point}

\item{mu}{An average}

\item{s}{A standard deviation}
}
\description{
Calculate a z-score for numeric data
}
