% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-piracy.R
\docType{data}
\name{piracy}
\alias{piracy}
\title{Piracy.}
\format{
A data frame with 534 rows and 8 variables.
\describe{
\item{name}{Name of member of Congress.}
\item{party}{Political party; (D)emocrat, (R)epublican, (I)ndependent.}
\item{state}{State represented.}
\item{money_pro}{Money from lobby groups in favor of the legislation.}
\item{money_con}{Money from lobby groups against the legislation.}
\item{years}{Number of years served in Congress.}
\item{stance}{Stance on legislation; yes, leaning yes, no, leaning no, unknown.}
\item{chamber}{Chamber of congress the member served in; house, sentate.}
}
}
\usage{
piracy
}
\description{
Data set containing monetary campaign donations received
by member of the US Congress in relation to anti-piracy
legislation that was on the table in the early 2000's.
}
\keyword{datasets}
