% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-project-data.R
\docType{data}
\name{project_data}
\alias{project_data}
\title{Project Data.}
\format{
A data frame with 8840 rows and 8 variables.
\describe{
\item{country}{Name of the country.}
\item{continent}{Name of the continent the country is on.}
\item{year}{Ranges from 1979-2018.}
\item{life_exp}{Life expectancy at birth.}
\item{net_users}{Number of people who use the internet.}
\item{non_net_users}{Number of people who do not use the internet.}
\item{population}{Number of people in the country.}
\item{num_phones}{Number of active cell phones in the country.}
}
}
\source{
\href{https://gapminder.org/data}{Gapminder}
}
\usage{
project_data
}
\description{
A data set compiled from multiple data sets available at gapminder.org/data.
}
\keyword{datasets}
