% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_ci.R
\name{prop_ci}
\alias{prop_ci}
\title{Calculate a confidence interval for categorical data}
\usage{
prop_ci(p_hat, alpha, n)
}
\arguments{
\item{p_hat}{A sample proportion}

\item{alpha}{A significance level}

\item{n}{Sample size}
}
\description{
Calculate a confidence interval for categorical data
}
